files/requirements.txt
Authlib==1.7.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
cryptography==47.0.0
Flask==3.1.3
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.5.0
idna==3.13
itsdangerous==2.2.0
Jinja2==3.1.6
joserfc==1.6.4
MarkupSafe==3.0.3
numpy==2.4.4
pandas==3.0.2
pillow==12.2.0
pycparser==3.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
qrcode==8.2
requests==2.33.1
six==1.17.0
SQLAlchemy==2.0.49
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
Werkzeug==3.1.8
